pyproject.toml
[project]
name = "final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dvc[s3]>=3.67.0",
    "kagglehub>=1.0.0",
    "numpy>=2.0.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=3.0.0",
    "torch>=2.0.0",
    "transformers>=4.51.0",
    "zarr>=3.1.6",
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu126", marker = "sys_platform == 'linux'" },
]
